alabproject/requirements.txt
Django==4.2.7
djangorestframework_simplejwt==5.3.0
drf_yasg==1.21.7
